+++ registrar/apps/core/management/commands/run_consumer.py
@@ -1,10 +1,13 @@
 """ Management command to run worker that will act on messages  """
 import logging
 
+from django.conf import settings
 from django.contrib.auth.models import Group
 from django.core.management.base import BaseCommand, CommandError
+from kombu import Connection
+from kombu.utils.debug import setup_logging
 
-from registrar.consumer import run_consumer_worker
+from registrar.apps.core.consumers import ProgramConsumer
 
 logger = logging.getLogger(__name__)
 
@@ -15,4 +18,11 @@ class Command(BaseCommand):
     help = 'Runs a worker to act on messages received from queue.'
 
     def handle(self, *args, **options):
-        run_consumer_worker()
+        setup_logging(loglevel='DEBUG')
+
+        with Connection(settings.BROKER_URL) as conn:
+         try:
+             ProgramConsumer(conn).run()
+         except KeyboardInterrupt:
+             print('bye bye')
+
